Added States

+++ src/main/java/assignment5/IStateStats.java
@@ -2,8 +2,8 @@ package assignment5;
 
 public interface IStateStats {
 
-    double demStat();
-    double repStat();
-    double electoralStat();
+    int demStat();
+    int repStat();
+    int electoralStat();
 
 }
